style(pad paragraph): add padding above the lorem paragraph
This is an effort to make sure the navbar does not cover any of the lorem ipsum paragraph when scrolled all the way to the top of the page. I am unsure of the best practice ways to accomplish this task. In this instance, a couple paragraphs were used, with '&nbsp;' inside, which adds non-breaking space. Non-breaking space is similar to just adding a blank paragraph, but it does not wrap and does not collapse.