feat(form icon): add a form input with an icon
This is another way to use the svg icons. This used to be done using input groups such as input-group-prepend and input-group-append to decipher between which side of the input the icon should reside. This is no longer used and bootstrap now just uses input-group-text and relies on the placement with regards to the form-control input.